Move Schedule.hpp to eclipse/EclipseState
Internal Schedule headers are not moved.
