+++ opm/parser/eclipse/IntegrationTests/TransMultIntegrationTests.cpp
@@ -24,7 +24,7 @@
 #include <opm/parser/eclipse/Parser.hpp>
 #include <opm/parser/eclipse/Parser/ParseContext.hpp>
 #include <opm/parser/eclipse/EclipseState.hpp>
-#include <opm/parser/eclipse/EclipseState/Schedule/Schedule.hpp>
+#include <opm/parser/eclipse/EclipseState/Schedule.hpp>
 
 using namespace Opm;
