+++ opm/parser/eclipse/Parser/tests/ParseContextTests.cpp
@@ -34,7 +34,7 @@
 #include <opm/parser/eclipse/Parser/ParseContext.hpp>
 
 #include <opm/parser/eclipse/EclipseState/Grid/EclipseGrid.hpp>
-#include <opm/parser/eclipse/EclipseState/Schedule/Schedule.hpp>
+#include <opm/parser/eclipse/EclipseState/Schedule.hpp>
 
 using namespace Opm;
